+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [Core] Added the default creation of volumes for containerd, etcd and root, if not indicated in the keoscluster
 * [Core] Support k8s v1.30
 * [Core] Deprecated Kubernetes versions prior to 1.28
+* [PLT-817] Bump Tigera Operator version to v3.28.2
 
 ## 0.17.0-0.5.2 (2024-06-25)